requirements.txt
-i https://pypi.org/simple
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.18.0
altair==4.2.2
anyascii==0.3.1
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cohere==3.2.6
comm==0.1.2
contourpy==1.0.7
coqpit==0.0.17
cycler==0.11.0
cython==0.29.28
datasets==2.9.0
dateparser==1.1.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg==1.4
filelock==3.9.0
flask==2.2.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
g2pkk==0.1.2
gitdb==4.0.10
gitpython==3.1.30
gruut[de]==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
huggingface-hub==0.12.0
idna==3.4
importlib-metadata==6.0.0
inflect==5.6.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.18.2
jieba==0.42.1
jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonlines==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.2
jupyter-core==5.2.0
jupyter-events==0.5.0
jupyter-server==2.2.1
jupyter-server-fileid==0.6.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.6.1
jupyter-ydoc==0.2.2
jupyterlab==3.6.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.19.0
kiwisolver==1.4.4
librosa==0.8.0
llvmlite==0.38.1
markdown-it-py==2.1.0
markupsafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mecab-python3==1.0.5
mistune==2.0.4
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8.1
notebook==6.5.2
notebook-shim==0.2.2
num2words==0.5.12
numba==0.55.2
numpy==1.22.4
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
pillow==9.4.0
platformdirs==2.6.2
pooch==1.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-readability-metrics==1.4.5
pyarrow==11.0.0
pycparser==2.21
pydantic[dotenv]==1.10.4
pydeck==0.8.0
pygments==2.14.0
pympler==1.0.1
pynndescent==0.5.8
pyparsing==3.0.9
pypinyin==0.48.0
pyrsistent==0.19.3
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyyaml==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
resampy==0.4.2
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
scikit-learn==1.2.1
scipy==1.10.0
semver==2.13.0
send2trash==1.8.0
setuptools==67.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.11.0
soupsieve==2.3.2.post1
stack-data==0.6.2
streamlit==1.17.0
tensorboardx==2.5.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
tornado==6.2
tqdm==4.64.1
trainer==0.0.20
traitlets==5.9.0
tts==0.10.2
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.1
unidic-lite==1.0.8
uri-template==1.2.0
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
werkzeug==2.2.2
xxhash==3.2.0
y-py==0.5.5
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.12.0
